Consolidate Internal to Utils
