+++ Source/Thirdweb/Private/ThirdwebUtils.cpp
@@ -2,22 +2,24 @@
 
 #include "ThirdwebUtils.h"
 
+#include <random>
+
+#include "HttpModule.h"
 #include "Thirdweb.h"
 #include "ThirdwebLog.h"
 #include "ThirdwebRuntimeSettings.h"
-
 #include "Dom/JsonObject.h"
-
 #include "GenericPlatform/GenericPlatformHttp.h"
-
 #include "Interfaces/IHttpRequest.h"
-
+#include "Interfaces/IPluginManager.h"
+#include "Kismet/GameplayStatics.h"
 #include "Kismet/KismetStringLibrary.h"
-
 #include "Policies/CondensedJsonPrintPolicy.h"
-
 #include "Serialization/JsonReader.h"
 #include "Serialization/JsonWriter.h"
+#include "Wallets/ThirdwebInAppWalletHandle.h"
+#include "Wallets/ThirdwebSmartWalletHandle.h"
+#include "Wallets/ThirdwebWalletHandle.h"
 
 #define LOCTEXT_NAMESPACE "Thirdweb"
 
@@ -109,6 +111,81 @@ namespace ThirdwebUtils
 				ANSI_TO_TCHAR(reinterpret_cast<const char*>(Request->GetContent().GetData()))
 			)
 		}
+
+		FString GetPluginVersion()
+		{
+			if (const TSharedPtr<IPlugin> Plugin = IPluginManager::Get().FindPlugin(TEXT("Thirdweb")); Plugin.IsValid())
+			{
+				return Plugin->GetDescriptor().VersionName;
+			}
+			return "0.0.0";
+		}
+		
+		// https://stackoverflow.com/a/58467162/12204515
+		FString GenerateUUID()
+		{
+			static std::random_device Device;
+			static std::mt19937 RNG(Device());
+
+			std::uniform_int_distribution Distribution(0, 15);
+
+			// ReSharper disable once CppTooWideScope
+			const char* Options = "0123456789abcdef";
+			constexpr bool Dash[] = {false, false, false, false, true, false, true, false, true, false, true, false, false, false, false, false};
+
+			std::string Result;
+			for (int i = 0; i < 16; i++)
+			{
+				if (Dash[i]) Result += "-";
+				Result += Options[Distribution(RNG)];
+				Result += Options[Distribution(RNG)];
+			}
+			return Result.c_str();
+		}
+		
+		// ReSharper disable CppPassValueParameterByConstReference
+		void SendConnectEvent(const FWalletHandle Wallet)
+		{
+			if (!IsInGameThread())
+			{
+				// Retry on the GameThread.
+				const FWalletHandle WalletCopy;
+				FFunctionGraphTask::CreateAndDispatchWhenReady([WalletCopy]()
+				{
+					SendConnectEvent(WalletCopy);
+				}, TStatId(), nullptr, ENamedThreads::GameThread);
+				return;
+			}
+			const UThirdwebRuntimeSettings* Settings = UThirdwebRuntimeSettings::Get();
+			if (!UThirdwebRuntimeSettings::AnalyticsEnabled() || UThirdwebRuntimeSettings::GetBundleId().IsEmpty() || UThirdwebRuntimeSettings::GetClientId().IsEmpty())
+			{
+				return;
+			}
+			FHttpModule& HttpModule = FHttpModule::Get();
+			const TSharedRef<IHttpRequest> Request = HttpModule.CreateRequest();
+			Request->SetVerb("POST");
+			Request->SetURL("https://c.thirdweb.com/event");
+			Request->SetHeader("Content-Type", "application/json");
+			Request->SetHeader("x-sdk-name", "UnrealEngineSDK");
+			Request->SetHeader("x-sdk-os", UGameplayStatics::GetPlatformName());
+			Request->SetHeader("x-sdk-platform", "unreal-engine");
+			Request->SetHeader("x-sdk-version", ThirdwebUtils::Internal::GetPluginVersion());
+			Request->SetHeader("x-client-id", Settings->ClientId);
+			Request->SetHeader("x-bundle-id", Settings->BundleId);
+			Request->SetTimeout(5.0f);
+
+			// ReSharper disable once CppLocalVariableMayBeConst
+			TSharedPtr<FJsonObject> JsonObject = MakeShareable(new FJsonObject);
+			JsonObject->SetStringField(TEXT("source"), TEXT("connectWallet"));
+			JsonObject->SetStringField(TEXT("action"), TEXT("connect"));
+			JsonObject->SetStringField(TEXT("walletAddress"), Wallet.ToAddress());
+			JsonObject->SetStringField(TEXT("walletType"), Wallet.GetTypeString());
+			Request->SetContentAsString(ThirdwebUtils::Json::ToString(JsonObject));
+			Request->ProcessRequest();
+		}
+		void SendConnectEvent(const FInAppWalletHandle Wallet) { SendConnectEvent(FWalletHandle(Wallet)); }
+		void SendConnectEvent(const FSmartWalletHandle Wallet) { SendConnectEvent(FWalletHandle(Wallet)); }
+		// ReSharper restore CppPassValueParameterByConstReference
 	}
 
 	namespace Maps
