+++ Source/Thirdweb/Public/ThirdwebUtils.h
@@ -2,6 +2,9 @@
 
 #pragma once
 
+struct FWalletHandle;
+struct FInAppWalletHandle;
+struct FSmartWalletHandle;
 class FJsonObject;
 class FJsonValue;
 class IHttpRequest;
@@ -86,6 +89,13 @@ namespace ThirdwebUtils
 		extern TArray<FString> MaskSensitiveString(const TArray<FString>& InStrings, const FString& MatchString, const FString& MaskCharacter = TEXT("*"), const int32 ShowBeginCount = 4, const int32 ShowEndCount = 4);
 		extern TArray<FString> MaskSensitiveString(const TArray<FString>& InStrings, const TArray<FString>& MatchStrings, const FString& MaskCharacter = TEXT("*"), const int32 ShowBeginCount = 4, const int32 ShowEndCount = 4);
 		extern void LogRequest(const TSharedRef<IHttpRequest>& Request, const TArray<FString>& SensitiveStrings = {});
+		
+		static FString GetPluginVersion();
+		static FString GenerateUUID();
+
+		static void SendConnectEvent(const FWalletHandle Wallet);
+		static void SendConnectEvent(const FInAppWalletHandle Wallet);
+		static void SendConnectEvent(const FSmartWalletHandle Wallet);
 	}
 }
